+++ src/controllers/ApiKeysController.ts
@@ -98,7 +98,7 @@ class ApiKeysController {
 
     const apiKeyId = request.params.id;
     if (!apiKeyId) {
-      response.status(400).send({ error: 'No API key\'s ID was provided' });
+      response.status(400).send({ error: "No API key's ID was provided" });
       return;
     }
 
@@ -111,7 +111,7 @@ class ApiKeysController {
     // Check if the current user is the owner of project
     const project = await Project.findOne({ _id: apiKey.projectId, ownerId: tokenResult.token.userId });
     if (!project) {
-      response.status(403).send({ error: 'You don\'t own this project' });
+      response.status(403).send({ error: 'You do not own this project' });
     }
 
     await apiKey.remove();
@@ -128,7 +128,7 @@ class ApiKeysController {
 
     const apiKeyId = request.params.id;
     if (!apiKeyId) {
-      response.status(400).send({ error: 'No API key\'s ID was provided' });
+      response.status(400).send({ error: "No API key's ID was provided" });
       return;
     }
 
@@ -141,7 +141,7 @@ class ApiKeysController {
     // Check if the current user is the owner of project
     const project = await Project.findOne({ _id: apiKey.projectId, ownerId: tokenResult.token.userId });
     if (!project) {
-      response.status(403).send({ error: 'You don\'t own this project' });
+      response.status(403).send({ error: 'You do not own this project' });
     }
 
     const { description, expiresAt } = request.body;